Extract height calculations to a hook
[#172103928]
